% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CRCLabel}
\alias{CRCLabel}
\title{CRC labels}
\format{
A factor vector of length \eqn{N}.
}
\usage{
CRCLabel
}
\description{
A factor vector containing the class labels corresponding to each sample
in \code{CRCFeature}. Each element indicates the diagnostic or molecular
subtype of the CRC sample.
}
\examples{
data(CRCLabel)
table(CRCLabel)
}
\seealso{
\code{\link{CRCFeature}}, \code{\link{LoadData}}
}
\keyword{datasets}
