% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_embeddings.R
\name{load_embeddings}
\alias{load_embeddings}
\title{Load in Histological Features}
\usage{
load_embeddings(feature, label = NULL, label_col = "label")
}
\arguments{
\item{feature}{The file path for the input data, the data should be
in csv, tsv format}

\item{label}{The file path for the input data, the data should be
in csv, tsv format}

\item{label_col}{The file path for the input data, the data should be
in csv, tsv format}
}
\value{
Returns a dataframe with feature columns and one label column
}
\description{
A function that load in histological features extracted from any
pathology foundation model. Optionally include labels
given log-likelihood, number of clusters, dimensionality of the
dataset and the number of observations.
}
\examples{
data(train_embeddings)
data(train_labels)
train_set <- load_embeddings(feature=train_embeddings, label=train_labels)

}
\references{
Schwarz, G. (1978). Estimating the dimension of a model. \emph{The Annals of Statistics}
6.

Akaike, H. (1973). Information theory and an extension of the maximum likelihood
principle. In \emph{Second International Symposium on Information Theory}, New York, NY,
USA, pp. 267–281. Springer Verlag.
}
