% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{LoadData}
\alias{LoadData}
\title{Calculates Information Criteria}
\usage{
LoadData(filepath, label = NULL)
}
\arguments{
\item{datapath}{The file path for the input data, the data should be
in csv, tsv format}
}
\value{
Returns an S3 object of class InfCriteria with results.
\itemize{
\item BICresults - A value of class numeric indicating BIC value.
\item AIC results - A value of class numeric indicating AIC value.
}
}
\description{
A function that calculates information criteria, BIC (Bayesian
Information Criterion) and AIC (Akaike Information Criterion)
given log-likelihood, number of clusters, dimensionality of the
dataset and the number of observations.
}
\examples{
HistData <- LoadData(filepath)
# Access BIC value
HistData$feature
HistData$label

}
\references{
Schwarz, G. (1978). Estimating the dimension of a model. \emph{The Annals of Statistics}
6.

Akaike, H. (1973). Information theory and an extension of the maximum likelihood
principle. In \emph{Second International Symposium on Information Theory}, New York, NY,
USA, pp. 267–281. Springer Verlag.
}
