% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_labels}
\alias{test_labels}
\title{Test labels (example data)}
\format{
A factor or character vector like \code{train_labels}.
}
\usage{
data(test_labels)
}
\description{
Test labels (example data)
}
\keyword{datasets}
