% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_embeddings}
\alias{train_embeddings}
\title{Training embeddings (example data)}
\format{
A data.frame with \eqn{n} rows (samples) and \eqn{p} columns (features).
}
\usage{
data(train_embeddings)
}
\description{
Training embeddings (example data)
}
\keyword{datasets}
