% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_embeddings}
\alias{visualize_embeddings}
\title{Visualize Embeddings Using Reduced Dimensions}
\usage{
visualize_embeddings(input_data, dimensions = 2, type = "pca")
}
\arguments{
\item{input_data}{An object of class \code{"histofeature"} created by
\code{\link{load_embeddings}}.}

\item{dimensions}{A positive integer specifying the number of dimensions
to visualize (between 2 and 8). Defaults to \code{2}.}

\item{type}{The dimensionality reduction method to use. Currently only
\code{"pca"} is supported.}
}
\value{
A \code{ggplot} object (for 2D visualization) or a
\code{ggmatrix} object (for multi-dimensional paired plots), displaying
the projected embeddings colored by label.
}
\description{
Generate a 2D scatterplot or a multi-dimensional paired plot of histological
feature embeddings using dimensionality reduction. Currently supports
Principal Component Analysis (PCA) for projection.
}
\note{
Documentation and debugging processes were refined with the assistance
of ChatGPT-5 to improve accuracy and clarity.
}
\examples{
data(train_embeddings)
data(train_labels)
train_set <- load_embeddings(feature=train_embeddings, label=train_labels)
visualize_embeddings(train_set)

}
\references{
R Core Team (2025). \emph{R: A Language and Environment for
Statistical Computing}. R Foundation for Statistical Computing,
Vienna, Austria. \url{https://www.R-project.org/}.

H. Wickham. ggplot2: Elegant Graphics for Data Analysis.
Springer-Verlag New York, 2016.

Schloerke B, Cook D, Larmarange J, Briatte F, Marbach M, Thoen E,
Elberg A, Crowley J (2025). \emph{GGally: Extension to 'ggplot2'}.
doi:10.32614/CRAN.package.GGally
\url{https://doi.org/10.32614/CRAN.package.GGally}, R package
version 2.4.0, \url{https://CRAN.R-project.org/package=GGally}.
}
