% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate model}
\usage{
evaluate_model(trained_model, test_data)
}
\arguments{
\item{trained_model}{A negative value of class numeric indicating
the log-likelihood.}

\item{test_data}{A positive integer indicating the number of clusters.}
}
\value{
a png
}
\description{
A function that calculates information criteria, BIC (Bayesian
Information Criterion) and AIC (Akaike Information Criterion)
given log-likelihood, number of clusters, dimensionality of the
dataset and the number of observations.
}
\examples{
data(train_embeddings)
data(train_labels)
data(test_embeddings)
data(test_labels)
train_set <- load_embeddings(feature=train_embeddings, label=train_labels)
test_set <- load_embeddings(feature=test_embeddings, label=test_labels)
model <- train_model(feature_embedding=train_set,
                      dr="pca", dr_k=20, model = "knn")
evaluate_model(model, test_set)
}
\references{
Schwarz, G. (1978). Estimating the dimension of a model. \emph{The Annals of Statistics}
6.

Akaike, H. (1973). Information theory and an extension of the maximum likelihood
principle. In \emph{Second International Symposium on Information Theory}, New York, NY,
USA, pp. 267–281. Springer Verlag.
}
