% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadData.R
\name{LoadData}
\alias{LoadData}
\title{Load in Histological Features}
\usage{
LoadData(feature, label = NULL, label_col = "label")
}
\arguments{
\item{feature}{The file path for the input data, the data should be
in csv, tsv format}

\item{label}{The file path for the input data, the data should be
in csv, tsv format}

\item{label_col}{The file path for the input data, the data should be
in csv, tsv format}
}
\value{
Returns a dataframe with feature columns and one label column
}
\description{
A function that load in histological features extracted from any
pathology foundation model. Optionally include labels
given log-likelihood, number of clusters, dimensionality of the
dataset and the number of observations.
}
\examples{
data(CRCFeature)
data(CRCLabel)
HistData <- LoadData(CRCFeature, CRCLabel)

}
\references{
Schwarz, G. (1978). Estimating the dimension of a model. \emph{The Annals of Statistics}
6.

Akaike, H. (1973). Information theory and an extension of the maximum likelihood
principle. In \emph{Second International Symposium on Information Theory}, New York, NY,
USA, pp. 267–281. Springer Verlag.
}
