% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{reduce_dim}
\alias{reduce_dim}
\title{Dimension reduction helper}
\usage{
reduce_dim(X, method = "pca", k = 20)
}
\arguments{
\item{X}{numeric matrix}

\item{method}{currently only "pca"}

\item{k}{integer; number of components to keep}
}
\value{
list with reduced matrix and model
}
\description{
Dimension reduction helper
}
