% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_embeddings}
\alias{test_embeddings}
\title{Test embeddings (example data)}
\format{
A data.frame like \code{train_embeddings}.
}
\usage{
data(test_embeddings)
}
\description{
Test embeddings (example data)
}
\keyword{datasets}
