% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_knn}
\alias{train_knn}
\title{KNN training helper}
\usage{
train_knn(X, y)
}
\arguments{
\item{X}{matrix or data.frame of predictors}

\item{y}{factor of class labels}
}
\value{
list with best tuned model and fitted knn3 object
}
\description{
KNN training helper
}
\keyword{internal}
