% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_knn}
\alias{train_knn}
\title{Train a K-Nearest Neighbors (KNN) Classifier}
\usage{
train_knn(X, y)
}
\arguments{
\item{X}{Matrix of predictor variables.}

\item{y}{Factor of class labels corresponding to the rows of \code{X}.}
}
\value{
The best fitted model using the best K.
}
\description{
Internal helper function that trains and tunes a KNN model.
}
