% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{train_labels}
\alias{train_labels}
\title{Training labels (example data)}
\format{
A factor or character vector of length \eqn{n}.
}
\usage{
data(train_labels)
}
\description{
Training labels (example data)
}
\keyword{datasets}
