% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_model.R
\name{train_model}
\alias{train_model}
\title{Train a classifier on embeddings}
\usage{
train_model(feature_embedding, dr = "pca", dr_k = 20, model = "knn")
}
\arguments{
\item{feature_embedding}{input}

\item{dr}{now only "pca"}

\item{dr_k}{integer, target dimension for reduction}

\item{model}{now only knn}
}
\value{
a list with fitted model, reduced features, and the DR info
}
\description{
Train a classifier on embeddings
}
\examples{
data(train_embeddings)
data(train_labels)
train_set <- load_embeddings(feature=train_embeddings, label=train_labels)
model <- train_model(feature_embedding=train_set,
                      dr="pca", dr_k=20, model = "knn")
}
