% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_embeddings}
\alias{visualize_embeddings}
\title{Visualize Embeddings using Reduced Dimensions}
\usage{
visualize_embeddings(input_data, dimensions = 2, type = "pca")
}
\arguments{
\item{input_data}{an object of class \code{"histofeature"}}

\item{dimensions}{A A positive integer indicating the dimension to visualize,
ranging from 2 to 8. The default is \code{2}.}

\item{type}{The type of dimension reduction techniques to perform, currently
only support \code{"pca"}.}
}
\value{
A \code{ggplot} or \code{ggmatrix} object that visualized the relationship.
}
\description{
A function that creates a 2D scatterplot or a multi-dimensional paired plot
of embeddings.
}
\examples{
data(train_embeddings)
data(train_labels)
train_set <- load_embeddings(feature=train_embeddings, label=train_labels)
visualize_embeddings(train_set)

}
\references{
Schwarz, G. (1978). Estimating the dimension of a model. \emph{The Annals of Statistics}
6.

Akaike, H. (1973). Information theory and an extension of the maximum likelihood
principle. In \emph{Second International Symposium on Information Theory}, New York, NY,
USA, pp. 267–281. Springer Verlag.
}
