% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize.R
\name{visualize_embeddings}
\alias{visualize_embeddings}
\title{Visualize the command}
\usage{
visualize_embeddings(input_data, dimensions = 2, type = "pca")
}
\arguments{
\item{input_data}{A negative value of class numeric indicating
the log-likelihood.}

\item{dimensions}{A positive integer indicating the number of clusters.}

\item{type}{A positive integer indicating the dimensionality
of the dataset.}
}
\description{
A function that calculates information criteria, BIC (Bayesian
Information Criterion) and AIC (Akaike Information Criterion)
given log-likelihood, number of clusters, dimensionality of the
dataset and the number of observations.
}
\examples{
data(train_embeddings)
data(train_labels)
train_set <- load_embeddings(feature=train_embeddings, label=train_labels)
visualize_embeddings(train_set)

}
\references{
Schwarz, G. (1978). Estimating the dimension of a model. \emph{The Annals of Statistics}
6.

Akaike, H. (1973). Information theory and an extension of the maximum likelihood
principle. In \emph{Second International Symposium on Information Theory}, New York, NY,
USA, pp. 267–281. Springer Verlag.
}
